--- __posts/1.md
@@ -1,11 +1,0 @@
----
-categories:
-  - Development
-date: '2023-07-10'
-description: 설명을 적어주세요.
-title: Next.js로 블로그 만들기
----
-
-**_content_**
-
-Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essentially unchanged. It was popularised in the 1960s with the release of Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing software like Aldus PageMaker including versions of Lorem Ipsum.
