--- __posts/2.md
@@ -1,24 +1,0 @@
----
-categories:
-  - Development
-date: '2023-07-11'
-description: 2번 글입니다.
-title: 언제 만들지
----
-
-# 제목 1
-
-## 제목 2
-
-### 제목 3
-
-**content**
-
-- content 1
-- content 1
-
-> 인용문
-
-```javascript
-<div>code</div>
-```
